Update lib.rs
This code creates an instance of the AccountInfo struct where the data field is an Rc<RefCell<&mut [u8]>> object that enforces the length of the data to be 128 bytes. You can adjust the DATA_LEN constant to match the actual length of the data you want to store in the AccountInfo object.